perf(web): pause the tenants-overview poll while the tab is hidden
TenantsOverview refreshed the cross-tenant aggregate every 60 seconds whether or
not anyone could see it, so a backgrounded tab billed a request a minute for a
surface nobody was looking at.

The active-alerts poll already handled this. Its visibility handling is lifted
into pollWhileVisible and both callers now share it, rather than the overview
growing a second copy of the same timer.
